+++ model_navigator/framework_api/commands/convert/converters/ts2torchtrt.py
@@ -26,25 +26,22 @@ def _get_precision(precision, precision_mode):
     if precision_mode == TensorRTPrecisionMode.HIERARCHY:
         enabled_precisions = {
             TensorRTPrecision.FP32: [torch.float],
-            TensorRTPrecision.TF32: [torch.float],
             TensorRTPrecision.FP16: [torch.float, torch.half],
             TensorRTPrecision.INT8: [torch.float, torch.half, torch.int8],
         }[precision]
     elif precision_mode == TensorRTPrecisionMode.SINGLE:
         enabled_precisions = {
             TensorRTPrecision.FP32: [torch.float],
-            TensorRTPrecision.TF32: [torch.float],
             TensorRTPrecision.FP16: [torch.half],
             TensorRTPrecision.INT8: [torch.int8],
         }[precision]
     else:
         raise ValueError(
             f"Unsupported precision mode {precision_mode}. Only {TensorRTPrecisionMode.HIERARCHY} and "
-            "{TensorRTPrecisionMode.SINGLE} are allowed"
+            f"{TensorRTPrecisionMode.SINGLE} are allowed"
         )
     return {
         "enabled_precisions": enabled_precisions,
-        "disable_tf32": precision == TensorRTPrecision.FP32,
     }