+++ CHANGELOG.md
@@ -23,6 +23,7 @@ limitations under the License.
   - new: example for loading Navigator Package for reproducing the results
   - new: create reproducing script for correctness and performance steps
   - new: TrtexecRunner for correctness and performance tests with [trtexec](https://docs.nvidia.com/deeplearning/tensorrt/developer-guide/index.html#trtexec) tool
+  - new: removed `TF32` precision for TensorRT and use TF32 support by default for models with FP32 precision
   - change: rename of `onnx_runtimes` export function parameters to `runtimes`
   - fix: do not save TF-TRT models to the .nav package
   - fix: do not save TF-TRT models from the .nav package